pyproject.toml
[project]
name = "roma"
version = "0.1.0"
description = "ROMA: Recursive Open Meta-Agent Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # "dspy>=2.6.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "aiofiles>=24.0.0",
    "tavily-python>=0.5.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]
all = [
    "roma[dev]",
    "httpx>=0.27.0",
    "rich>=13.0.0",
]

[project.scripts]
roma = "src.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
